requirements.txt
Django==4.1.1
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-import-export==2.8.0
django-js-asset==2.0.0
django-redis==5.2.0
djangorestframework==3.13.1
Pillow==9.2.0